feat: 유저 예약 기능 수정
